backend/requirement.txt
Django==3.2.23
mysqlclient==2.2.4
numpy
pandas==2.0.3
Pillow==9.0.1
PyMySQL==1.1.0
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.19.3
scikit-learn==1.3.0
scipy==1.10.1
segmentation-models-pytorch==0.3.3
shap==0.42.1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0